create event deatils template
- move event details data to view module
