+++ src/view/form.js
@@ -1,4 +1,5 @@
 import {EVENT_TYPES} from '../consts.js';
+import {createEventDetails} from './event-details';
 
 const createEventTypeTemplate = (types) => {
   return types.map((type) => `
@@ -12,42 +13,12 @@ const createEventTypeTemplate = (types) => {
 const eventTo = createEventTypeTemplate(EVENT_TYPES.to);
 const eventIn = createEventTypeTemplate(EVENT_TYPES.in);
 
-const renderFormEventOffers = (options) => {
-  return options.map(({id, title, price}) => `
-    <div class="event__offer-selector">
-      <input class="event__offer-checkbox  visually-hidden" id="event-offer-${id}-1" type="checkbox" name="event-offer-${id}">
-      <label class="event__offer-label" for="event-offer-${id}-1">
-        <span class="event__offer-title">${title}</span>
-        &plus;
-        &euro;&nbsp;<span class="event__offer-price">${price}</span>
-      </label>
-    </div>
-  `).join(``);
-};
-
-const renderFormEventDescription = (description) => {
-  const {text, photos} = description;
-
-  return `
-    <p class="event__destination-description">${text}</p>
-
-    <div class="event__photos-container">
-      <div class="event__photos-tape">
-        ${photos.map((src) => `
-          <img class="event__photo" src="${src}" alt="Event photo">
-        `).join(``)}
-      </div>
-    </div>
-  `;
-};
-
 export const createTripFormTemplate = (obj) => {
   const {
     options,
     description,
   } = obj;
-  const eventOffers = renderFormEventOffers(options);
-  const eventDescription = renderFormEventDescription(description);
+  const eventDetails = createEventDetails(options, description);
 
   return (`
     <form class="trip-events__item  event  event--edit" action="#" method="post">
@@ -119,20 +90,7 @@ export const createTripFormTemplate = (obj) => {
           <span class="visually-hidden">Open event</span>
         </button>
       </header>
-      <section class="event__details">
-        <section class="event__section  event__section--offers">
-          <h3 class="event__section-title  event__section-title--offers">Offers</h3>
-
-          <div class="event__available-offers">
-            ${eventOffers}
-          </div>
-        </section>
-
-        <section class="event__section  event__section--destination">
-          <h3 class="event__section-title  event__section-title--destination">Destination</h3>
-          ${eventDescription}
-        </section>
-      </section>
+      ${eventDetails}
     </form>
   `);
 };
